Update View Charts/Report buttons behavior and clean up process chain code structure
